django-api/requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
simplejwt==2.0.1
sqlparse==0.5.1
typing==3.7.4.3
typing-extensions==4.12.2
